cineflix-app/agent/requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.0
httpx>=0.24.0
langchain-groq>=0.1.0
python-dotenv>=1.0.0
